feat: separated client and server examples, first client example
