+++ examples/docker-compose.yml
@@ -1,0 +1,9 @@
+services:
+  mcp-everything:
+    image: node:lts-alpine
+    command: ["npx", "-y", "@modelcontextprotocol/server-everything", "streamableHttp"]
+    ports:
+      - "3001:3001"
+    environment:
+      PORT: "3001"
+    restart: unless-stopped
