+++ examples/src/main/scala/chimp/server/adderMcp.scala
@@ -1,29 +1,24 @@
-//> using dep com.softwaremill.chimp::core:0.1.7
+//> using dep com.softwaremill.chimp::chimp-server:0.2.0
 //> using dep com.softwaremill.sttp.tapir::tapir-netty-server-sync:1.11.50
 //> using dep ch.qos.logback:logback-classic:1.5.20
 
-package chimp
+package chimp.server
 
-import chimp.server.*
 import io.circe.Codec
-import sttp.model.Header
 import sttp.tapir.*
 import sttp.tapir.server.netty.sync.NettySyncServer
 
-@main def mcpAuthApp(): Unit =
-  case class Input(a: Int, b: Int) derives Codec, Schema
+case class Input(a: Int, b: Int) derives Codec, Schema
 
+@main def mcpApp(): Unit =
   val adderTool = tool("adder")
     .description("Adds two numbers")
     .withAnnotations(ToolAnnotations(idempotentHint = Some(true)))
     .input[Input]
 
-  def logic(i: Input, headers: Seq[Header]): Either[String, String] =
-    val tokenMsg =
-      headers.find(_.name == "test_header").map(t => s"token: ${t.value} (header name used: ${t.name})").getOrElse("no token provided")
-    Right(s"The result is ${i.a + i.b} ($tokenMsg)")
+  def logic(i: Input): Either[String, String] = Right(s"The result is ${i.a + i.b}")
 
-  val adderServerTool = adderTool.handleWithHeaders(logic)
+  val adderServerTool = adderTool.handle(logic)
 
   val mcpServerEndpoint = mcpEndpoint(List(adderServerTool), List("mcp"))
